+++ bytecodemanipulation/data/shared/expressions/MacroParameterAcessExpression.py
@@ -31,7 +31,7 @@ class MacroParameterAccessExpression(AbstractAccessExpression):
 
             for instr in instructions:
                 if instr.has_local():
-                    instr.change_arg_value("OUTER_" + instr.arg_value)
+                    instr.change_arg_value(":" + instr.arg_value)
 
             return instructions
 
@@ -60,7 +60,7 @@ class MacroParameterAccessExpression(AbstractAccessExpression):
 
             for instr in instructions:
                 if instr.has_local():
-                    instr.change_arg_value("OUTER_" + instr.arg_value)
+                    instr.change_arg_value(":" + instr.arg_value)
 
             return instructions